Fix Google geocode result parsing to handle multiple result nodes in an XML response.
Google can return multiple results if you issue a geocoding request that is broad in nature, e.g. a street name rather than a fully qualified street address. This exposed a bug: XPath parsing of the XML was skipping past the first two Placemarks, which have no postal code, and returning the postal code from the third Placemark. The fix is to make the XPath searches relative to the current node, rather than to the root node.